requirements.txt
fastapi
uvicorn
openai
httpx
beautifulsoup4
requests
python-dotenv
psycopg[binary]
